+++ GenotypeFormat.py
@@ -27,7 +27,7 @@ def format_vcf_genotype(vcf_allele1_char, vcf_allele2_char, ploidy):
         vcf_allele1_char (string): 0,1,2 etc.
         vcf_allele2_char (string): 0,1,2, etc.
         vcf_record (vcf._Record): Record for the entry analyzed
-        ploidy(int): Expected ploidy. 
+        ploidy(int): Expected ploidy.
 
     Returns
         string: String representation of genotype (e.g., "0/1"), adjusted for haploid calls if applicable
@@ -76,7 +76,7 @@ def convert_indel_genotype_to_vcf(nucleotide_genotypes, vcf_record, is_deletion,
     Returns:
         string: VCF genotype (e.g, "0/1")
     """
-    if len(nucleotide_genotypes) == 0:
+    if not nucleotide_genotypes:
         return format_vcf_genotype(".", ".", ploidy)
 
     if len(nucleotide_genotypes) > 1:
@@ -163,7 +163,7 @@ class RecordCombiner(object):
         """
         alleles = set()
         for record in self._bpm_records:
-            alleles.update(record.get_plus_strand_alleles())
+            alleles.update(record.plus_strand_alleles)
         return list(combinations_with_replacement(alleles, 2))
 
     def _record_inconsistent_with_genotype(self, record, genotype):
@@ -182,7 +182,7 @@ class RecordCombiner(object):
         if record_int_genotype == 0:
             return False
 
-        plus_strand_alleles = record.get_plus_strand_alleles()
+        plus_strand_alleles = record.plus_strand_alleles
         record_plus_genotype = convert_ab_genotype_to_nucleotide(
             record_int_genotype, plus_strand_alleles)
 
@@ -290,7 +290,7 @@ class GenotypeFormat(object):
         return _Format(GenotypeFormat.get_id(), 1, "String", GenotypeFormat.get_description())
 
     def generate_sample_format_info(self, bpm_records, vcf_record, sample_name):
-        """ 
+        """
         Get the sample genotype
 
         Args:
@@ -311,7 +311,7 @@ class GenotypeFormat(object):
                 int_genotype = self._genotypes[record.index_num]
                 if int_genotype != 0:
                     nucleotide_genotypes.append(convert_ab_genotype_to_nucleotide(
-                        int_genotype, bpm_records[0].get_plus_strand_alleles()))
+                        int_genotype, bpm_records[0].plus_strand_alleles))
             vcf_genotype = convert_indel_genotype_to_vcf(
                 nucleotide_genotypes, vcf_record, bpm_records[0].is_deletion, ploidy)
         else:
@@ -328,7 +328,7 @@ class GenotypeFormat(object):
                     nucleotide_genotype = ('-', '-')
                 else:
                     nucleotide_genotype = convert_ab_genotype_to_nucleotide(
-                        sample_genotype, bpm_records[0].get_plus_strand_alleles())
+                        sample_genotype, bpm_records[0].plus_strand_alleles)
                 vcf_genotype = convert_nucleotide_genotype_to_vcf(
                     nucleotide_genotype, vcf_record, ploidy)